+++ controllers/operations.js
@@ -1,14 +1,12 @@
-
-const mongoose = require('mongoose')
-const operationSchema = mongoose.model('Operations')
 // const Operation = require('../model/Operation')
+const verifyToken = require('../webservice/tokenVerification.js')
 
 const userOperations = async (req, res) => {
   try {
-    const user = req.query.user
-
-    const operations = await operationSchema.find({ userEmail: user.toString() })
-    // const parsedOperations = new Operation(operations)
+    const verify = await verifyToken(req, res)
+    if (verify.message === 'Unauthorized' || verify.message === 'Invalid token') {
+      return res.status(verify.status).send(verify.message)
+    }
 
     res.status(201).send([])
   } catch (error) {