+++ jest.config.js
@@ -3,5 +3,6 @@ module.exports = {
   testPathIgnorePatterns: [
     'test/end-to-end/'
     // Add more patterns for other ignored tests
-  ]
+  ],
+  setupFiles: ['<rootDir>/jest.env']
 }