+++ routes/user.js
@@ -2,7 +2,8 @@ const express = require('express')
 const router = express.Router()
 const userControllers = require('../controllers/user.js')
 
-// router.get('/users', userControllers.getUsers)
 router.post('/register', userControllers.register)
+router.post('/login', userControllers.login)
+router.get('/users', userControllers.getUsers)
 
 module.exports = router