+++ server.js
@@ -3,19 +3,21 @@ require('dotenv').config()
 const express = require('express')
 const app = express()
 require('./persistence/database')
+require('./aspects')
+
 const apiRoutes = require('./routes/apiRoutes')
 const cors = require('cors')
 const swaggerUi = require('swagger-ui-express')
 const swaggerDocument = require('./swagger/swagger.json')
-
-const sampleDataSet = require('./test/dataset/test-dataset')
+const sampleDataSet = require('./dataset/init-dataset')
+const periodicCotizations = require('./service/internal/periodicCotizations')
 
 const corsOptions = {
   origin: process.env.BASE_URL
 }
 
 // Sample dataset
-if (process.env.ENVIRONMENT === 'production') sampleDataSet.init().then(console.log('Data Loaded')).catch(console.log('Data Failed'))
+if (process.env.ENVIRONMENT === 'development') sampleDataSet.init().then(console.log('Data Loaded')).catch(console.log('Data Failed'))
 
 // Api
 app.use(express.urlencoded({ extended: true }))
@@ -33,9 +35,10 @@ app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument))
 app.use(cors(corsOptions))
 app.use('/api', apiRoutes)
 
-// app.listen(PORT, console.log(`App is running in ${process.env.NODE_ENV} mode on port ${PORT}`))
 const port = process.env.PORT || 3000
-app.listen(port)
-console.log('App listening in port ' + port)
+app.listen(port, () => {
+  console.log('Server running on port ' + port)
+  periodicCotizations.startAgenda()
+})
 
 module.exports = app