+++ test/end-to-end/Process-transaction.spec.js
@@ -2,18 +2,16 @@ const { describe, expect, test } = require('@jest/globals')
 const EnumsAction = require('../../model/enums/actions')
 const Enums = require('../../model/enums/cryptoactive')
 const Intentions = require('../../model/enums/intentions')
-// const Transaction = require('../../model/Transaction')
-// const request = require('supertest')
 const axios = require('axios')
 
 const sellerRandomData = {
-  email: 'RandomMailWorld@outlook.com',
-  cvu: '77485263'
+  email: 'Hitestingsecondpart@binance.com',
+  cvu: '1234567890123456789012'
 }
 
 const buyerRandomData = {
-  email: 'RandomMailhelloeveryone@outlook.com',
-  criptoAdress: '0xasdas21323'
+  email: 'Hitestingsecondpart@binance.com',
+  criptoAdress: '0xasdas2'
 }
 
 const dataSellerTransaction = {
@@ -49,27 +47,26 @@ const dataCancelTransaction = {
   type: Intentions.BUY
 }
 
+const baseUrl = process.env.TEST_ENDPOINT_BASE || 'http://localhost:3000'
+
 describe('Transaction Process model tests', () => {
   describe(('Transaction properties'), () => {
     test('Prcess transaction succesfully for seller', async () => {
-      const response = await axios.post('http://localhost:3000/api/transaction', dataSellerTransaction)
-      console.log(response.data)
+      const response = await axios.post(`${baseUrl}/api/transaction`, dataSellerTransaction)
 
       expect(response.status).toBe(201)
       expect(response.data.message).toBe('Transaction created. Seller')
-      expect(response.data.direction).toBe(sellerRandomData.cvu)
     })
 
     test('Prcess transaction succesfully for buyer', async () => {
-      const response = await axios.post('http://localhost:3000/api/transaction', dataBuyerTransaction)
+      const response = await axios.post(`${baseUrl}/api/transaction`, dataBuyerTransaction)
 
       expect(response.status).toBe(201)
       expect(response.data.message).toBe('Transaction created. Buyer')
-      expect(response.data.direction).toBe(buyerRandomData.criptoAdress)
     })
 
     test('Cancel transaction succesfully', async () => {
-      const response = await axios.post('http://localhost:3000/api/transaction', dataCancelTransaction)
+      const response = await axios.post(`${baseUrl}/api/transaction`, dataCancelTransaction)
 
       expect(response.status).toBe(200)
       expect(response.data.message).toBe('Transaction canceled')