+++ Changelog.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+## 2.0.0 (third tag)
+
+Features:
+  - Endpoints securized with JWT
+  - Architecture test
+  - env file for end-to-end tests.
+  - Redis DB added for some services.
+  - LOGS reflections added for endpoints
+  - All registered users endpoint added
+
+Fixes:
+- Sample data now set the data by endpoints instead directly saving it in DB
+- Fixed all not working endpoints
+- Production fake binance endpoint fixed
+- "production" env var behaviour swapped to "development"
+
 ## 1.1.0 (second tag)
 
 Features: